cleaned up settings loading, added error messages for ganache being installed, and generalised loading tests
