+++ utils/deployutils.py
@@ -2,6 +2,7 @@ import time
 
 from web3 import Web3, HTTPProvider
 from solc import compile_files
+from utils.generalutils import *
 
 BLOCKCHAIN_ADDRESS = "http://localhost:8545"
 W3 = Web3(HTTPProvider(BLOCKCHAIN_ADDRESS))
@@ -40,28 +41,6 @@ def fresh_accounts(num_accs):
         raise Exception("""W3.eth.accounts doesn't contain enough accounts,
                         restart ganache with more accounts (i.e. ganache-cli -a 500)""")
     return accs[:num_accs]
-
-
-class TERMCOLORS:
-    BLUE = '\033[94m'
-    GREEN = '\033[92m'
-    YELLOW = '\033[93m'
-    RED = '\033[91m'
-    BOLD = '\033[1m'
-    RESET = '\033[0m'
-
-
-def to_seconds(seconds=0, minutes=0, hours=0, days=0, weeks=0):
-    total_time = seconds
-    mult = 60
-    total_time += minutes*mult
-    mult *= 60
-    total_time += hours*mult
-    mult *= 24
-    total_time += days*mult
-    mult *= 7
-    total_time += weeks*mult
-    return total_time
 
 
 def attempt(function, func_args, init_string, print_status=True, print_exception=True):
@@ -103,14 +82,6 @@ def fast_forward(seconds=0, minutes=0, hours=0, days=0, weeks=0):
     W3.providers[0].make_request("evm_increaseTime", [total_time])
     force_mine_block()
 
-def take_snapshot():
-    x = W3.providers[0].make_request("evm_snapshot", [])
-    force_mine_block()
-    return x
-
-def restore_snapshot(snapshot):
-    W3.providers[0].make_request("evm_revert", [snapshot['result']])
-    force_mine_block()
 
 def take_snapshot():
     x = W3.providers[0].make_request("evm_snapshot", [])
